+++ event/state.go
@@ -170,6 +170,7 @@ type ModPolicyContent struct {
 	Recommendation string `json:"recommendation"`
 }
 
+// Deprecated: MSC2716 has been abandoned
 type InsertionMarkerContent struct {
 	InsertionID id.EventID `json:"org.matrix.msc2716.marker.insertion"`
 	Timestamp   int64      `json:"com.beeper.timestamp,omitempty"`