+++ requests.go
@@ -331,6 +331,7 @@ type ReqPutPushRule struct {
 	Pattern    string                     `json:"pattern"`
 }
 
+// Deprecated: MSC2716 was abandoned
 type ReqBatchSend struct {
 	PrevEventID id.EventID `json:"-"`
 	BatchID     id.BatchID `json:"-"`
@@ -340,6 +341,12 @@ type ReqBatchSend struct {
 
 	StateEventsAtStart []*event.Event `json:"state_events_at_start"`
 	Events             []*event.Event `json:"events"`
+}
+
+type ReqBeeperBatchSend struct {
+	Forward    bool           `json:"forward"`
+	MarkReadBy id.UserID      `json:"mark_read_by,omitempty"`
+	Events     []*event.Event `json:"events"`
 }
 
 type ReqSetReadMarkers struct {